[fix](unique function) fix in predicate extract non-constant for unique function (#55417)
InPredicateExtractNonConstant will extract non-constant options from in
predicate, for the purpose of be speed up evaluating the in predicate.

but for the sql `a + random() in (b, 10, 20, 30)`, it will rewrite as `a
+ random() = b or a + random() in (10, 20, 30)`, then BE will calculate
the random() twice for one row.

so if the in predicate's compare expression contains unique function,
don't extract it.