[refactor](nereids) agg strategy refactor (#54079)
### What problem does this PR solve?

#### Current Issues:
1. When the children of an agg satisfy a distribution, redundant
PhysicalAggregates are generated.
2. count(distinct a, b) generates a three-stage gather: the first two
stages deduplicate a and b, the third stage gathers the data, and then
counts. (This modification avoids a direct gather by adding a local agg
before the gather.)
3. The original code did not use statistical information to determine
the different Aggregate strategies. This refactor incorporates
statistical information as a basis for selection:
3.1 Utilizes statistical information in the logic for determining
whether to use multi_distinct or CTE splitting: When the ndv of the
group by key is low and the ndv of the distinct key is high,
multi_distinct is advantageous.
3.2 In the single distinct scenario, statistical information is also
used to determine whether to use a three-stage or four-stage Aggregate.
When the ndv of the group by key is low, a four-stage Aggregate is used.


#### REWRITE PHASE
##### DistinctAggStrategySelector
Add the DistinctAggStrategySelector rule to handle scenarios with
multiple distincts. Use statistical information to determine when to use
CTE splitting or multidistinct.
After this rule is implemented, only single distinct scenarios need to
be considered.

##### DistinctAggregateRewriter
Add the DistinctAggregateSplitter rule to handle scenarios with a single
distinct. Determine whether to split or use multi-distinct. Distinguish
between scenarios with and without group by.
No group by: Not processed here. With group by: If splitting is possible
and statistical information indicates that multi-stage splitting is more
optimal, split.

So, after this is complete, cases with distinct without group by:
No processing, move to the cascades phase. 
Cases with distinct with group by:
1. distinct is rewritten to multi-distinct,
2. it is split, leaving distinct.
3. if it is neither rewritten to multi-distinct nor split, then
splitting is performed directly into two, three, or four stages in the
cascades phase.

#### CASCADES STAGE
##### SplitAgg
Add a SplitAgg rule to handle scenarios without distinct.
1. Split the Agg into two layers: local and global.
This allows you to avoid splitting the Agg into two stages, for example,
in group_concat scenarios. (If not, shuffle first, then agg.)
2. This also implements a single-stage Agg.

##### SplitAggMultiPhaseWithoutGbyKey
ADD implementation rule: SplitAggMultiPhaseWithoutGbyKey
only process agg without group by key, and with one distinct function
provide there three rewrite:
```sql
select count(distinct a) from t
splitToThreePhase:
agg(count(a); distinct global)
  +--gather
    +--agg(count(a); distinct local)
      +--agg(group by a; global)
        +--hashShuffle(a)
splitToFourPhase:
agg(count(a); distinct global)
  +--gather
    +--agg(count(a); distinct local)
      +--agg(group by a; global)
        +--hashShuffle(a)
          +--agg(group by a; local)
twoPhaseAggregateWithFinalMultiDistinct:
agg(sum0(c1))
  +--gather
    +--agg(multi_distinct(a) as c1)
      +--hashShuffle(a)
```
splitToThreePhase and twoPhaseAggregateWithFinalMultiDistinct provide
plan when scan(or other plan) satisfy distributed expr

##### SplitAggMultiPhase
ADD implementation rule: SplitAggMultiPhase
only process agg with group by key, and with one distinct function
select count(distinct a) group by b (deduplicated agg hashShuffle by
group by key b)
```sql
splitToTwoPlusOnePhase:
  agg(group by b, count(a); distinct global)
    +--agg(group by a,b; global)
      +--hashShuffle(b)
        +--agg(group by a,b; local)
  agg(group by b, count(a); distinct global)
    +--agg(group by a,b; global)
      +--hashShuffle(b)
splitToTwoPlusTwoPhase:
  agg(group by b, count(a); distinct global)
    +--hashShuffle(b)
      +--agg(group by b, count(a); distinct local)
        +--agg(group by a,b; global)
          +--hashShuffle(a,b)
            +--agg(group by a,b; local)
  agg(group by b, count(a); distinct global)
    +--hashShuffle(b)
      +--agg(group by b, count(a); distinct local)
        +--agg(group by a,b; global)
          +--hashShuffle(a,b)
splitToOnePlusTwoPhase: (deduplicated agg hashShuffle by distinct key a)
  agg(group by b, count(a); distinct global)
    +--hashShuffle(b)
      +--agg(group by b, count(a); distinct local)
        +--agg(group by a,b; global)
          +--hashShuffle(a)
```
The second plan of splitToTwoPlusOnePhase and splitToTwoPlusTwoPhase
provide plan when scan(or other plan) satisfy group by distributed expr;
splitToOnePlusTwoPhase provide plan when scan(or other plan) satisfy
distinct key distribution.

#### RequestPropertyDeriver
Implement visitPhysicalHashAggregate. The main logic is:
If the request received by the AGG is hash-distributed with the
distribution column set S1, and the request sent by the AGG is also
hash-distributed with the distribution column set S2,
and S = S2.intersect(S1), and the NDV of S1 is calculated, and if the
NDV of S1 is found to be no less than a threshold, then the AGG sends
the child a hash distribution based on S1.
Otherwise, the AGG sends the child a hash distribution based on S.
When directly generating a three-stage/four-stage AGG, the partitionby
columns are specified. If the partitionby columns are not empty, the
child is hash-distributed based on the partitionby columns.

#### ChildrenPropertiesRegulator
Disabled one-stage AGG: AGG-Distribute
Special handling for one-stage AGG with CTE:
AGG-Distribute-CTE.
If the Distribute option is Gather, the option is disabled.
If the Distribute option is HashShuffle, the option is determined based
on statistics.
    If no statistics are available, the option is enabled.
If statistics are available but the shuffle column is skewed, the option
is disabled.