[feat](unique function) add project for unique function (#48449)
### What problem does this PR solve?

If an unique function exists multiple times, then be will calculate it
multiple times for each row, so it will be error.

for example: `filter(random() between 10 and 20)`, after rewrite the
`between`, it will get `filter(random() >= 10 and random() <= 20)`, this
will contains two random in one expression, the two RANDOM should be
one, so we add a project, then we can get `filter(k >= 10 and k <= 20)
-> project(random() as k)`

this PR also fix BETWEEN expression bug introduced by #55407
- between shouldn't be PropagateNullable because FoldConstantRuleOnFE
will rewrite a propagate nullable expression to null if any children is
NULL, but for sql `10 between null and 5` should be `FALSE`, not `NULL`;
- after analyzed between expression, it will get an AND expression, then
anlyzed join other conjunctions, need to extract conjuncts of each
analyzed other conjunction (that is flattern AND).