[fix](unique function) fix virtual slot with unique function (#55716)
if a sub expression exists multiple time in filter, rule
PushDownVirtualColumnsIntoOlapScan will extract it as a virtual slot and
push it into olap scan.

but if this sub expression itself contains an unique function which
exists multiple times, then it should not push into olap scan, rule
AddProjectForUniqueFunction will extract the unique function for it.

But we can't construct a real example SQL satisfy the above case, maybe
it's not exists in real SQL, we just still fix it for safe reason. And
this PR we don't add a real SQL regression test, we just test it with fe
ut.