[fix](mysql protocol)Set more stmt exists flag correctly when forward to master. (#55711)
When execute multiple statements in one batch and
CLIENT_MULTI_STATEMENTS is set, Doris server need to set
SERVER_MORE_RESULTS_EXISTS flag in the return packet before the last
statement. But when the Observer forward stmt to Master, this
SERVER_MORE_RESULTS_EXISTS is not set, cause the following statements
failed to execute.
This pr forward a boolean value to Master, so the Master FE knows it is
the last statement or not, and could set SERVER_MORE_RESULTS_EXISTS
correctly.