[improve](nereids) Adjust the conditions for confirming hot values (#55472)
### What problem does this PR solve?
In the current version, we determine whether a value is a hot value
based on its occurrence proportion. However, this method gives rise to
two issues:

1. If the number of rows in the table is small, even if a value appears
only once, it will still be identified as a hot value — which increases
the overhead of statistical information derivation.
2. If the number of rows in the table is large (e.g., 10 million) and
the ndv (Number of Distinct Values) is also large (e.g., 10,000), the
average number of occurrences per value is 1000. In this case, if a
value accounts for 10% of the total, its actual occurrences will reach 1
million — 1000 times the average. If such a value is not estimated as a
hot value, it will result in a 1000-fold error.

By the way, adjust the value of the hot ratio representing a proportion
from a percentage format (e.g., 10 represents 10%) to a decimal format
(e.g., 0.1 represents 10%).

Therefore, we revise the criteria for identifying hot values to:
Actual number of occurrences > Average number of occurrences ×
hot_value_threshold

where hot_value_threshold is a session variable with a default value of
10.