[fix](cast) Fix the incorrect nullable property setting for variant cast to JSON. (#55899)
In non-strict mode, the nullable property for variant cast to JSON
should be set to true.
Note, however, that the code for variant cast to JSON is not yet fully
implemented—in non-strict mode, it currently casts variant to string
first, then from string to JSON.
In strict mode, it will still throw an error.