[fix](exchange) remove const_cast in VDataStreamMgr (#56088)
### What problem does this PR solve?

Previously there was a const_cast here, but in our internal tests this
occasionally caused a hard-to-reproduce core dump.
We suspect it was caused by the const_cast, so we switched to making a
copy here.
In fact, for PBlock, most of the data resides in the PColumnMeta
column_metas field, so the copy overhead is small.
To make the intent explicit, we do not use
```C++
std::unique_ptr<PBlock> pblock_ptr = std::make_unique<PBlock>(request->blocks(i));
```

### Release note

None

### Check List (For Author)

- Test <!-- At least one of them must be included. -->
    - [ ] Regression test
    - [ ] Unit Test
    - [ ] Manual test (add detailed scripts or steps below)
    - [ ] No need to test or manual test. Explain why:
- [ ] This is a refactor/code format and no logic has been changed.
        - [ ] Previous test can cover this change.
        - [ ] No code files have been changed.
        - [ ] Other reason <!-- Add your reason?  -->

- Behavior changed:
    - [ ] No.
    - [ ] Yes. <!-- Explain the behavior change -->

- Does this need documentation?
    - [ ] No.
- [ ] Yes. <!-- Add document PR link here. eg:
https://github.com/apache/doris-website/pull/1214 -->

### Check List (For Reviewer who merge this PR)

- [ ] Confirm the release note
- [ ] Confirm test cases
- [ ] Confirm document
- [ ] Add branch pick label <!-- Add branch pick label that this PR
should merge into -->