[improve](nereids) if a value occurs many times in a column, take it as hot value (#55827)
### What problem does this PR solve?

In the previous PR #55472, whether a value qualifies as a hot value for a certain column depends on whether its occurrence frequency is higher than the average frequency of all values (i.e., 1/ndv). However, this approach is not conducive to the use of skew join; therefore, we also include values from columns with uniform distribution but small ndv into the collection of hot values.