[fix] (agg_strategy) fix result wrong when the multi_distinct_func and count distinct multi expr exists same time (#56271)
### What problem does this PR solve?

Related PR: #54079

Problem Summary:
1. Added a check for multi_distinct_count(a,b) in the MultiDistinctCount
constructor to prevent the use of multiple columns. Because BE doesn't
report an error in this case, it only uses the first argument of the
multi_distinct function, resulting in incorrect results.
2. In scenarios without a group by key, when multi_distinct_func and
count(distinct a,b) appear together, the original code converts
count(distinct a,b) to multi_distinct_count(a), resulting in incorrect
results. The correct approach is to use multi-stage splitting when count
distinct multi_expr appears.
3. Removed the mustUseMultiDistinct flag. This flag is useless.