[Improve](segment) Put Segment footer into index page cache
This commit modified the caching logic for the Segment
footer.

Previously, the Segment footer was cached in the `DATA_PAGE`
cache. This was incorrect as the footer is metadata, not
data. This commit changes it to use the more appropriate
`INDEX_PAGE` cache.

Additionally, the calculation for `_meta_mem_usage` has been
adjusted for better accuracy. The footer size is no longer
included in the calculation since it's now managed by the
`StoragePageCache`. The memory estimation for column readers
is also updated to use
`config::max_segment_partial_column_cache_size` for a more
realistic value.