[opt](nereids) optimize bs downgrade alg. (#56279)
### What problem does this PR solve?
PR#36784: Bucket shuffle join shall be disabled if the number of scan
instances for the left table of the join is less than 10.
The value "10" should not be hardcoded here; instead, it should be
calculated based on the cluster size. This PR will replace the fixed
value "10" with a dynamic value equal to 0.8 times the maximum
concurrency of the cluster.