[feat](case when) replace nested inner case when duplicate condition to true/false (#56469)
for nested case when, replace the inner case duplicate condition to
true/false when this condition also exists in outer case condition:

1. if it exists in outer case's current branch condition, replace it with TRUE:
case when A then (case when A and B then 1 else 2 end)
 ...
end

then inner case condition A will replace with TRUE:
case when A then  (case when TRUE and B then 1 else 2 end)
...
end

2. if it exists in outer case's previous branch condition, replace it with FALSE:
case when A then C
     when B then (case when A and D then 1 else 2 end)
 ...
end

then inner case condition A will replace with FALSE:
case when A then C
     when B then (case when FALSE and D then 1 else 2 end)
...
end

this PR also opt fold case when and fold if statement. for case when /
if expression, if all their branches values equals, then rewrite them to
the same value.