[fix](string) fix string len check of nested types (#56963)
Fix bug of checking string length of complex type.
When doing substring to truncate nested string column for complex types,
should not use `block->get_by_position(slot_index)`, because the column
returned by this method is the original column of complex type, not the
string column that we want to do substring.