[refactor](insert) unified INSERT_JOB to INSERT for EtlJobType (#56780)
NSERT_JOB was introduced in https://github.com/apache/doris/pull/26146,
the purpose is to specify the job id for insert job. But there is no
need to add a new job type, which would make the code difficult to
understand. This pr aim to unified INSERT_JOB to INSERT.