[feature](mtmv)create mtmv can use view (#56423)
The main changes are as follows:
- If the MTMV is in a schema change state, recheck whether the
materialized view's SQL statement meets the requirements before
refreshing, and check whether the PCT table has changed.
- Add a data structure in the StatementContext to store the tables and
views directly depended on in the plan.
- Add a data structure to store the relationship between materialized
views and views.
- The location for adding filter conditions during partition refresh has
been moved to CheckPolicy, as the previous location could only access
the parsed plan.
- Clear the snapshot information of the materialized view when a view is
deleted or modified.