[fix](nereids)set RuntimeFilterInfo only on BE which is merge node (#57108)
set RuntimeFilterInfo only on BE which is merge node.
note: RuntimeFilterInfo only contains runtime-filter merge information,
not runtime filter descriptor
In the previous pull request #56978, we moved the runtime info (which
records how runtime filters are merged) from the instance level to the
Backend (BE) level. In a multi-BE environment, when a runtime filter
needs to be merged, the BE does not perform the merge using the
mergeInstance specified by the Runtime Filter descriptor. Instead, it
selects a BE that has the RuntimeInfo set to perform the merge. This
causes BEs that should not be responsible for merging to wait for other
nodes to send local runtime filters, resulting in a time