[feat](nereids) opt condition expression in filter / join / if /case when (#57025)
For a condition expression, replace null to false, null safe equal to
equal. The condition include filter condition, join condition, if
condition, case when condition.

And for the condition expression, only replace those sub expression
which its ancestors to the condition root are all AND / OR / CASE WHEN /
IF. So, for a expression  in a filter, the first
null can be replaced, but the second null cann't be replaced because its
parent is NOT, not in AND / OR / CASE WHEN / IF.

For a expression  in condition, if one of them is not-nullable,
then it can rewrite to . Note that if a expression is not a
condition,  can rewrite to  require that both x and y
are not-nullable.