[feat](nereids) opt condition expression in filter / join / if /case when (#57025)
For a condition expression, replace null to false, null safe equal to
equal. The condition include filter condition, join condition, if
condition, case when condition.

And for the condition expression, only replace those sub expression
which its ancestors to the condition root are all AND / OR / CASE WHEN /
IF. So, for a expression `null and not(null)` in a filter, the first
null can be replaced, but the second null cann't be replaced because its
parent is NOT, not in AND / OR / CASE WHEN / IF.

For a expression `x <=> y` in condition, if one of them is not-nullable,
then it can rewrite to `x = y`. Note that if a expression is not a
condition, `x <=> y` can rewrite to `x = y` require that both x and y
are not-nullable.

Here is some example:

1. replace null with false

```
null  and (null  or null) and not(null)   =>  false and (false or false) and not(null)

case when null and not (null) then ...   =>   case when false and not (null) then ...

if(null and not(null), ...) =>  if (false and not(null), ...)
```

2. replace null safe equal to equal if at least one of its child is
non-nullable

```
x  <=>  3 and not(x <=> 3)   => x = 3 and not(x <=> 3)

case when x <=> 3 then ...    =>  case when x = 3 then ...

if(x <=> 3, ...)  => if(x=3, ...)
```

3. rewrite if expression in condition if its else branch is true / false
literal.

suppose the if is in a filter, then will have
```
 if(c, p, true)  => not(c <=> true) or p

if(c, p, false)  => c and p
```

Currently, the expression rewrite framework cann't known the
expression's plan, In order to indicate that the expression is on the
filter / join, we add the expression's plan and source to the
ExpressionRewriteContext to indicate the expression's location.