[opt](nereids) print slotReference in explain verbose (#57167)
### What problem does this PR solve?
In the result of explain verbose, most of columns of SlotDescriptors are
null. It is not easy to trace column in TupleDescriptor table. In this
pr, we add caption for SlotDescriptor. if column is null, the caption is
its corresponding slotreference.toString().

for example : "select sum(a) as mysum from t;"
we have "SlotDescriptor{id=5, col=mysum#3, ..."