[fix](external catalog) Distinguish between functions used to refresh cache and change catalog properties (#56639)
The `resetToUninitialized` method should only be called when altering catalog properties, as it resets the entire catalog to an uninitialized state. Previously, `resetToUninitialized` was mistakenly called in `refreshMgr`. Refresh should only refresh the cache, not the entire catalog. Resetting the entire catalog could cause certain clients to shut down during a cache refresh, leading to errors.

This PR mainly changes:
1. Change the `onRefreshCache()` method
    This method will invalid db/table names cache and all meta cache
2. For `refresh` and `hms listener`, only call `onRefreshCache()`
3. For `alter catalog`, call `resetToUninitialized()`

TODO:
There is still a problem when we `alter catalog`, the ongoing task may be failed or stucked.