[Improve](job) add history task info in task tvf show (#57361)
### What problem does this PR solve?

Currently, a job only retains the most recent task execution
information, and when it pauses or reports an error, the task
information is cleared. However, this lacks observability.
To this end, we've added a history task display.
Note：This information is only stored in memory and is cleared after a
restart.