[fix](maxcompute)fix maxcompute push down predicates null pointer. (#57474)
### What problem does this PR solve?
Problem Summary:
This PR is a continuation of PR #55635. Based on the current
investigation, the conclusion in PR #55635 appears to be incorrect — the
reason why the predicate column cannot be found in the columnMap remains
unclear.
To help with further debugging, this PR adds additional log information.

In addition, during the `refresh table` process, the
MaxComputeExternalTable object remains unchanged, so the objects it
holds also stay the same. When initializing the table schema, the map
maintained by this object may have repeated `put` operations. So, move
this map to `MaxComputeSchemaCacheValue`, which better matches its
lifecycle and avoids repeated updates during table refresh.