[refine](function) Remove unused code from the function framework and optimize the default NULL-handling logic. (#57811)
There used to be many unused parameters in the function framework; this
PR removes them.
The previous default-NULL handling framework inserted intermediate
results into the block and then erased them — when blocks are large this
caused significant overhead.
Now intermediates are handled using a temporary block, and a
simple_insert function is provided that only inserts values into the
data.