[chore] (streamload) add enable stream load profile config (#57919)
When streamload encounters slow performance issues, we need to enable
the stream load profile.
However, current implementations require users to actively enable it by
setting the stream load HTTP header. On the one hand, the cost of
communicating and configuring with users is relatively high; on the
other hand, the time for problem location will also be extended.
Therefore, a configuration on the FE side is added to dynamically enable
profiling.

Co-authored-by: zhangshixin.1024 <zhangshixin.1024@bytedance.com>