[feat](case when) Push upper function into case when branch (#56941)
push upper function into case when branch.

for expression f with n arguments a1, a2, ..., an, if one of its
argument is case when/if/nvl/nullif,
and the others are literals, then we can push f into each branch of case
when/if/nvl/nullif.
The rewrite rule is as follows:

```sql
f(a1, a2, ..., (case when c1 then p1 when c2 then p2 else p3 end), ..., an)

 => case when c1 then f(a1, a2, ..., p1, ..., an)
          when c2 then f(a1, a2, ..., p2, ..., an)
          else  f(a1, a2, ..., p3, ..., an) end
```

for example:

```
2 > case when c1 then 1 
               when c2 then 2
               else 3
        end
```

will rewrite to 

```
case when c1 then true
        when c2 then false
        else false
end
```